feat: Add sse handler
